fix: Decimal → float em dashboard_schemas (recharts exige número, não string)

+++ backend/src/schemas/dashboard_schemas.py
@@ -1,52 +1,51 @@
 import datetime
-from decimal import Decimal
 
 from pydantic import BaseModel
 
 
 class HoraBucket(BaseModel):
     hora: int
-    faturamento: Decimal
+    faturamento: float
 
 
 class ProdutoTop(BaseModel):
     item_id: int
     nome: str
     quantidade: int
-    faturamento: Decimal
+    faturamento: float
 
 
 class DiaFaturamento(BaseModel):
     data: datetime.date
-    faturamento: Decimal
+    faturamento: float
 
 
 class ComandaAbertaItem(BaseModel):
     id: int
     identificacao: str
     qtd_itens: int
-    total: Decimal
+    total: float
     aberta_ha_minutos: int
 
 
 class DashboardHistoricoItem(BaseModel):
     data: datetime.date
-    faturamento: Decimal
-    total_compras: Decimal
+    faturamento: float
+    total_compras: float
 
 
 class DashboardResumoAnualItem(BaseModel):
     mes: int
-    faturamento: Decimal
-    total_compras: Decimal
+    faturamento: float
+    total_compras: float
 
 
 class DashboardResponse(BaseModel):
-    faturamento_hoje: Decimal
-    ticket_medio_hoje: Decimal
+    faturamento_hoje: float
+    ticket_medio_hoje: float
     comandas_abertas: int
     comandas_fechadas_hoje: int
-    lucro_estimado_hoje: Decimal
+    lucro_estimado_hoje: float
     faturamento_por_hora: list[HoraBucket]
     top_10_produtos: list[ProdutoTop]
     ultimos_30_dias: list[DiaFaturamento]
